refactor(ui): extract MessageDivider + text/reasoning/compaction renderers
Slice 11b.1 step 9 of the message-part.tsx split. Pull the
non-tool part renderers (text, reasoning, compaction) and the
MessageDivider used both by them and by session-turn.tsx's
turn-change banner into a focused module.

New: packages/ui/src/components/message-part-core-renderers.tsx
  - MessageDivider          divider with centered label
  - PART_MAPPING["compaction"]
  - PART_MAPPING["text"]
  - PART_MAPPING["reasoning"]

MessageDivider is re-exported from message-part.tsx so the existing
session-turn.tsx import path stays valid; importing the new module is
required as a side-effect for the registry assignments.

tsgo --noEmit GREEN. bun test: 478 pass / 3 pre-existing fail.
